refactor: 수량 컴포넌트 공통화
