+++ README.md
@@ -30,7 +30,7 @@
 ## step4
 
 - [x] step3 리뷰 반영
-- [ ] 수량 컴포넌트 공통화
+- [x] 수량 컴포넌트 공통화
 - [ ] 상품 목록 수량 컴포넌트 추가
 - [ ] 상품 목록 수량 관련 비즈니스 로직 추가
 - [ ] 테스트 코드 확인
